requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

# add packages 
torch
torchvision
tensorboard
scikit-image
matplotlib
numpy
pandas
PySide6
yapf